+++ src/main/java/spring/chess/score/Calculatable.java
@@ -1,6 +1,6 @@
 package spring.chess.score;
 
-import spring.chess.game.ChessSet;
+import spring.chess.player.ChessSet;
 
 public interface Calculatable {
     Score calculate(ChessSet chessSet);
