+++ src/test/java/chess/game/ChessSetTest.java
@@ -1,15 +1,15 @@
 package chess.game;
 
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import spring.chess.board.ChessBoard;
 import spring.chess.board.ChessBoardCreater;
 import spring.chess.location.Location;
 import spring.chess.piece.type.Pawn;
 import spring.chess.piece.type.Piece;
+import spring.chess.player.ChessSet;
 import spring.chess.score.Score;
 import spring.chess.team.Team;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import spring.chess.game.ChessSet;
 
 import java.util.HashMap;
 import java.util.NoSuchElementException;
