+++ src/test/java/chess/player/PlayerTest.java
@@ -1,15 +1,15 @@
 package chess.player;
 
-import spring.chess.board.ChessBoard;
-import spring.chess.board.ChessBoardCreater;
-import spring.chess.game.ChessSet;
-import spring.chess.location.Location;
-import spring.chess.piece.type.Piece;
-import spring.chess.score.Score;
-import spring.chess.team.Team;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import spring.chess.board.ChessBoard;
+import spring.chess.board.ChessBoardCreater;
+import spring.chess.location.Location;
+import spring.chess.piece.type.Piece;
+import spring.chess.player.ChessSet;
 import spring.chess.player.Player;
+import spring.chess.score.Score;
+import spring.chess.team.Team;
 
 import java.util.Collections;
 import java.util.Map;
