+++ src/test/java/chess/score/PawnReduceScoreCalculableTest.java
@@ -1,17 +1,17 @@
 package chess.score;
 
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import spring.chess.board.ChessBoard;
 import spring.chess.board.ChessBoardCreater;
-import spring.chess.game.ChessSet;
 import spring.chess.location.Col;
 import spring.chess.location.Location;
 import spring.chess.location.Row;
 import spring.chess.piece.type.Piece;
+import spring.chess.player.ChessSet;
 import spring.chess.score.PawnReduceScoreCalculable;
 import spring.chess.score.Score;
 import spring.chess.team.Team;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 
 import java.util.Map;
