+++ src/test/java/chess/score/ScoreCalculatorTest.java
@@ -1,15 +1,15 @@
 package chess.score;
 
-import spring.chess.board.ChessBoard;
-import spring.chess.board.ChessBoardCreater;
-import spring.chess.game.ChessSet;
-import spring.chess.location.Location;
-import spring.chess.piece.type.Piece;
-import spring.chess.score.Score;
-import spring.chess.team.Team;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import spring.chess.board.ChessBoard;
+import spring.chess.board.ChessBoardCreater;
+import spring.chess.location.Location;
+import spring.chess.piece.type.Piece;
+import spring.chess.player.ChessSet;
+import spring.chess.score.Score;
 import spring.chess.score.ScoreCalculator;
+import spring.chess.team.Team;
 
 import java.util.Map;
