+++ src/test/java/chess/score/ScoreCalcultorTest.java
@@ -1,15 +1,15 @@
 package chess.score;
 
+import org.junit.jupiter.api.Test;
 import spring.chess.board.ChessBoard;
 import spring.chess.board.ChessBoardCreater;
-import spring.chess.game.ChessSet;
-import org.junit.jupiter.api.Test;
+import spring.chess.player.ChessSet;
 import spring.chess.score.Calculatable;
 import spring.chess.score.Score;
 import spring.chess.score.ScoreCalculator;
 
-import static spring.chess.team.Team.WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
+import static spring.chess.team.Team.WHITE;
 
 class ScoreCalcultorTest {
     @Test
